fix: avoid resetting store state when registering a dynamic module
Fixes https://github.com/vuejs/vuex/issues/2197

At the moment, when registering a dynamic module, we call
`resetStoreState()` just to register the getters for the new module.

It seems unnecessary to reset the entire store state in this case, and
this actually also leads to [other issues][1].

This change is based on the test case added in
https://github.com/vuejs/vuex/pull/2201

The approach taken in this change is to refactor the getter registration
into its own function, and call that new method when registering a
dynamic module instead of resetting the store state.

[1]: https://github.com/vuejs/vuex/issues/2197